PythonSDK_offers/pyproject.toml
[tool.poetry]
name = "offers"
version = "0.1.0"
description = "Async Python SDK for Offers API"
authors = ["Radek Pastyrik: radek.pastyrik@gmail.com"]
license = "MIT"
readme = "README.md"
packages = [{ include = "offers_sdk" }]
keywords = ["sdk", "async", "httpx", "offers"]

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.25"
pydantic = "^2.0"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
respx = "^0.20"
ruff = "^0.4"
mypy = "^1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
